move hash functions to crypto, move constants
